Changed parse_output_capability_report(...) to return fuel mix format
The parse_output_capability_report(...) function now returns a list of dictionaries with the fuel mix format matching that of the get_generation(...) function. The idea is that parse_output_capability_report(...) may be called multiple times when stitching multiple days' reports together. I'm not super happy with its structure currently (feels like doing three things in one function). However, I'm putting off refactoring the function until I get a feel for how to encapsulate the functionality given the get_generation(...) function's other options.

I got a start on handling the 'latest' parameter in the get_generation(...) function, but it currently doesn't trim the results. The behaviour of the function does not currently match the description of the base client interface.