Added Adequacy Report parser to be used for forecast data.
The _parse_adequacy_report(...) method has been implemented an is unit tested. However, it's not referenced by forecast conditions in the get_generation(...) method yet. The Adequacy Report highlighted that the IESO has an 'Other' fuel type that is not used in other reports but maps to WattTime pyiso fuels.

There also exists a 'Dispatchable Load' fuel type that has no mapping in the WattTime pyiso fuels. Is it most appropriate to dtop it or combine it with the 'Other' fuel appropriately?