#119 Fixed erroneous calls to self.utcify(...) with datetime parameter.
The method expects to receive a timestamp string as a parameter, and converts
it to a timezone-aware datetime in UTC. I was passing in timezone-naive
datetime values, but being caught gracefully within the method. This fixes
those erroneous calls and clarifies some parameter names internally.

While implementing the change I found that
GeneratorOutputByFuelHourlyReportHandler was not obeying min_datetime
and max_datetime parameters when appending results (i.e. start_at and
end_at). This commit fixes that erroneous behaviour.