#130 Increased default request timeout to 30 seconds.
I've increased the default request timeout to 30 seconds because
test_get_lmp_dataframe_fifteen(...) from test_caiso.py was
sporadically failing for me, not because the request was bad but
because processing time was right at the threshold of the request
processing timeout.

Increasing the default timeout from 20 seconds to 30 seconds seemed
reasonable, since 20 seconds is rather short. However, any further
increases to the timeout due to large .zip requests from CAISO should
possibly be isolated to the CAISOClient.