Fix CAISO unit/integration test and then separate them (#130) (#149)
* #130 Fixed test_dt_index(...) in test_caiso.py

The DatetimeIndex class was moved to from pandas.tseries.index to
pandas.core.indexes.datetimes with the release of 0.20.1:

http://pandas.pydata.org/pandas-docs/stable/whatsnew.html#reorganization-of-the-library-privacy-changes

This means that our requirement of pandas >= 0.18 could cause
test_dt_index(...) to fail in test_caiso.py sporadically if users fulfill
the requirement with pandas >= 0.20.1. It felt foolish to increment one of
 pyiso's core requirements for the sake of a unit test. So I changed the test
 to check for the type(...) class name rather than the type(...) exactly.

* #130 Increased default request timeout to 30 seconds.

I've increased the default request timeout to 30 seconds because
test_get_lmp_dataframe_fifteen(...) from test_caiso.py was
sporadically failing for me, not because the request was bad but
because processing time was right at the threshold of the request
processing timeout.

Increasing the default timeout from 20 seconds to 30 seconds seemed
reasonable, since 20 seconds is rather short. However, any further
increases to the timeout due to large .zip requests from CAISO should
possibly be isolated to the CAISOClient.

* Removed TIMEOUT_SECONDS constant from BaseClient.

The field was not being referenced anywhere. It only caused confusion
given that __init__ has a timeout_seconds parameter with a default value.

* #130 Split up integration tests and unit tests out of test_caiso.py

They now follow the directory conventions being established with #130.

* #130 Increased LMP request timeout for CAISO to 45 seconds.

* #130 Pulled static content out of test_caiso.py into fixtures directory.

* #130 Rely on setUp(...) to call client_factory('CAISO') before each test.

Unit and integration tests were each creating their own CAISOClient at that
start of each test. That type of setup can be left to the setUp(...) method.

* #130 Fixed BaseClient test timeout.

All we should be checking is that the request timeout is assignable. Checking
its default value with a unit test is just annoying.

* #130 Fixed byte encoding issues in mocked response.

This was causing errors in Python 3.