Increment version to 0.4.0 to reflect breaking change.
Commits 22323c6, 79464e1, and fa7524c remove the get_lmp(...)
method from BaseClient and increment the version to 0.3.20. This
is a backwards-incompatible change which we should give a
greater indication of via the package version number.

The industry conventions of semantic versioning (http://semver.org)
suggest that the first version number position should be
incremented for backward-incompatible API changes. That would
mean pyiso should be incremented to 1.0.0. That feels a bit odd, so
I suggest that we at least increment the version number to 0.4.0,
which is reflected in this commit.