Fixes read_excel(parse_cols=[...]) bug starting with pandas 0.21
The continuous integration builds have been failing since c015878. This is no fault of our own, but panas 0.21 was released around that same time and the build server is using it.

The parse_cols argument was deprecated in 0.21 in favour of usecols.

https://pandas.pydata.org/pandas-docs/version/0.21/whatsnew.html#deprecations

Though it's supposed to be backward compatible, something is broken in the unit tests. This commit changes the kwarg that the unit test is using to usecols. Although read_excel(...) in pandas < 0.21 does not document the argument (e.g. http://pandas.pydata.org/pandas-docs/version/0.20/generated/pandas.read_excel.html vs. http://pandas.pydata.org/pandas-docs/version/0.21/generated/pandas.read_excel.html) the kwarg is used internally when ._parse_excel(...) instantiates TextParser which calls _handle_usecols(...). It all works.

I'm hesitant to bump requirements.txt and setup.py to pandas>=0.21 just to fix this. I'll just allow older versions of pandas to handle the usecols kwarg in an undocumented fashion (as described above) unless this turns up as a problem elsewhere.