#156 Replace references to static responses in test_trade.py
The TestNYISOTrade class in test_trade.py was mocking static responses in an integration test. With improvements made to mocked responses in #183, testing static responses are handled by unit tests.

This commit changes the integration tests in TestNYISOTrade to use current times, so that the tests dynamically test our ability to integration with the external NYISO server (in line with #130).