#182 Pared down assertions in typical standard time / daylight saving time tests
They now only check timestamps and are not sensitive to the order of fuels as they are parsed.

This fixes an error with ordering that fuels are parsed in Python 3, but was missed in #182. Since the unit tests were continuously failing on the CI server, I only noticed now now (since they're working again).